Extract TextBox Git ignore pipe handling
