+++ .github/swift-file-length-budget.tsv
@@ -104,7 +104,7 @@
 944	Sources/CommandPalette/CommandPaletteSettingsToggle.swift
 943	Sources/Cloud/VMClient.swift
 942	Packages/iOS/CmuxMobileShellUI/Tests/CmuxMobileShellUITests/WorkspaceMacSelectionTests.swift
-937	Sources/TextBoxMentionIndexStore.swift
+850	Sources/TextBoxMentionIndexStore.swift
 929	Sources/Panels/TerminalPanel.swift
 918	Sources/Panels/BrowserPopupWindowController.swift
 905	Sources/CmuxSSHURLRequest.swift
